Remove Cosmopolitan ifdefs
If you want to use miniaudio with cosmopolitan, it's recommended that
you use the "cosmoaudio" API, which is a wrapper library provided by
cosmopolitan libc that lets you use the speaker and microphone. It
causes miniaudio.h to be included as a zip file in your ape binary.
That'll be extracted at runtime and built with the host compiler.

You can still compile miniaudio.h with cosmocc. But that's only useful
for accessing its pure libraries for audio file encoding and decoding.
What you should do in that case, is create a .c file like this:

    #include "stb_vorbis.h"
    #define MA_NO_DEVICE_IO
    #define MA_NO_RUNTIME_LINKING
    #define MINIAUDIO_IMPLEMENTATION
    #include "miniaudio.h"

Which can then be compiled and statically linked with cosmocc.